Added User Review and View models

+++ TerziyskiClima.Data/Models/Product.cs
@@ -7,12 +7,12 @@ namespace TerziyskiClima.Data.Models
     public class Product
     {
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string Description { get; set; }
         public int Id { get; set; }
         public virtual ICollection<Purchase> Purchases { get; set; }
 
-        public Product(string name, decimal price, string description)
+        public Product(string name, double price, string description)
         {
             Name = name;
             Price = price;
