+++ TerziyskiClima.Services/Interfaces/IProductService.cs
@@ -7,8 +7,8 @@ namespace TerziyskiClima.Services.Interfaces
 {
     public interface IProductService
     {
-        public Product Add(string name, decimal price, string description);
-        public void Edit(int id, string name, decimal price, string description);
+        public Product Add(string name, double price, string description);
+        public void Edit(int id, string name, double price, string description);
         public void Remove(int id);
         public Product GetProductById(int id);
         public List<Product> GetProducts();
